requirements.txt
streamlit>=1.31.0
requests>=2.31.0
python-dotenv>=1.0.0
rich>=13.9.4
markdown-it-py>=3.0.0
pygments>=2.19.1
mdurl>=0.1.2
